Support-Webhook: Remove redundant webhook route
